# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GoTasks is a robust API for managing company tasks and collaborations.\nIt provides comprehensive features for:\n- **Company and Collaborator Management**: Onboard companies and manage staff with hierarchical roles.\n- **Activity and Task Lifecycle Tracking**: Create and monitor activities and tasks through various stages.\n- **Role-Based Access Control (RBAC)**: Secure endpoints with granular permissions for Admins, Ops, Owners, and Managers.\n- **Performance Ratings and Assignments**: Assign tasks to collaborators and rate their performance.\n- **Secure Authentication**: JWT-based authentication for secure access.",
        "title": "GoTasks API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/company/activity/tasks/{activity_id}/{id}/perform": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Perform a task by updating its status for current collaborator",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task | Company Manager or Owner"
                ],
                "summary": "Perform task",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update payload",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_task.TaskUpdateStatusDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_MessageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/company/manager-owner/activity/tasks/{activity_id}/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List tasks by activity ID with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task | Company Manager or Owner"
                ],
                "summary": "List tasks by activity",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of items to skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Maximum number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "title"
                        ],
                        "type": "string",
                        "default": "id",
                        "description": "Field to sort by",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "default": "ASC",
                        "description": "Sort order",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter expression",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_Paginated-internal_task_ActivityTaskDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new task within an activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task | Company Manager or Owner"
                ],
                "summary": "Create task",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task create payload",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_task.TaskCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-internal_task_TaskDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/company/manager-owner/activity/tasks/{activity_id}/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a single task by ID within an activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task | Company Manager or Owner"
                ],
                "summary": "Get task",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-internal_task_ActivityTaskDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a task by ID within an activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task | Company Manager or Owner"
                ],
                "summary": "Update task",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task update payload",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_task.TaskUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-internal_task_TaskDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a task by ID within an activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task | Company Manager or Owner"
                ],
                "summary": "Delete task",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_MessageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/company/manager-owner/activity/tasks/{activity_id}/{id}/status": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the status of a task within an activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task | Company Manager or Owner"
                ],
                "summary": "Update task status",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update payload",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_task.TaskUpdateStatusDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_MessageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "go-tasks_internal_lifecycle.LifecycleStatusDTO": {
            "type": "string",
            "enum": [
                "CREATED",
                "IN_PROGRESS",
                "BLOCKED",
                "COMPLETED",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "LifecycleStatusCreatedDTO",
                "LifecycleStatusInProgressDTO",
                "LifecycleStatusBlockedDTO",
                "LifecycleStatusCompletedDTO",
                "LifecycleStatusCancelledDTO"
            ]
        },
        "go-tasks_internal_pkg_httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "go-tasks_internal_pkg_httpx.MessageDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation completed"
                }
            }
        },
...
        "go-tasks_internal_pkg_httpx.Paginated-internal_task_ActivityTaskDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_task.ActivityTaskDTO"
                    }
                }
            }
        },
        "go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_MessageDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/go-tasks_internal_pkg_httpx.MessageDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_Paginated-internal_task_ActivityTaskDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Paginated-internal_task_ActivityTaskDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "go-tasks_internal_pkg_httpx.Response-internal_task_ActivityTaskDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_task.ActivityTaskDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "go-tasks_internal_pkg_httpx.Response-internal_task_TaskDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_task.TaskDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "internal_task.ActivityTaskDTO": {
            "type": "object",
            "properties": {
                "activity_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lifecycle": {
                    "$ref": "#/definitions/internal_task.TaskLifecycleDTO"
                },
                "lifecycle_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_task.TaskCreateDTO": {
            "type": "object",
            "properties": {
                "lifecycle": {
                    "$ref": "#/definitions/internal_task.TaskLifecycleCreateDTO"
                },
                "title": {
                    "type": "string",
                    "example": "Prepare monthly report"
                }
            }
        },
...
        "internal_task.TaskLifecycleCreateDTO": {
            "type": "object",
            "properties": {
                "due_date": {
                    "type": "string",
                    "example": "2025-01-10T18:00:00Z"
                },
                "init_date": {
                    "type": "string",
                    "example": "2025-01-01T08:00:00Z"
                }
            }
        },
...
        "internal_task.TaskUpdateDTO": {
            "type": "object",
            "properties": {
                "lifecycle": {
                    "$ref": "#/definitions/internal_task.TaskLifecycleUpdateDTO"
                },
                "title": {
                    "type": "string",
                    "example": "Prepare monthly report (updated)"
                }
            }
        },
        "internal_task.TaskUpdateStatusDTO": {
            "type": "object",
            "properties": {
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/go-tasks_internal_lifecycle.LifecycleStatusDTO"
                        }
                    ],
                    "example": "IN_PROGRESS"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}